+++ tracker/utils/math.cpp
@@ -8,12 +8,12 @@ void set_flag(uint8_t &flag, const uint8_t bit, const bool value) {
     if (value) {
         flag |= bit;
     } else {
-        flag &= ~bit;
+        flag &= static_cast<uint8_t>(~bit);
     }
 }
 
 void combine_hash(size_t &seed, const size_t value) {
-    seed ^= value + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+    seed ^= value + 0x9e3779b9 + (seed << 6U) + (seed >> 2U);
 }
 
 void combine_hash(size_t &seed, const std::string &string) {
